Admin layouts & registered_by

+++ database/migrations/2024_07_01_210237_add-stripe-elements-to-users-table.php
@@ -1,5 +1,4 @@
 <?php
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -14,10 +13,13 @@ return new class extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('stripe_customer_id')->nullable();
-            $table->string('stripe_payment_method_id')->nullable();
+            if (!Schema::hasColumn('users', 'stripe_customer_id')) {
+                $table->string('stripe_customer_id')->nullable();
+            }
+            if (!Schema::hasColumn('users', 'stripe_payment_method_id')) {
+                $table->string('stripe_payment_method_id')->nullable();
+            }
         });
-        
     }
 
     /**
@@ -28,8 +30,12 @@ return new class extends Migration
     public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('stripe_customer_id');
-            $table->dropColumn('stripe_payment_method_id');
+            if (Schema::hasColumn('users', 'stripe_customer_id')) {
+                $table->dropColumn('stripe_customer_id');
+            }
+            if (Schema::hasColumn('users', 'stripe_payment_method_id')) {
+                $table->dropColumn('stripe_payment_method_id');
+            }
         });
     }
 };
